refactor: Moving components from HistoryLogGroup.tsx to ContributorAvatars.tsx and HistoryLogGroupEntries.tsx
