+++ src/library-authoring/generic/history-log/HistoryLogGroup.tsx
@@ -1,18 +1,18 @@
-import { ComponentProps, ReactNode, useState } from 'react';
+import { ReactNode } from 'react';
 import moment from 'moment';
 import classNames from 'classnames';
 import { Avatar, Collapsible, Icon, Stack, useToggle } from '@openedx/paragon';
 import { KeyboardArrowDown, KeyboardArrowUp } from '@openedx/paragon/icons';
-import { FormattedMessage, useIntl } from '@edx/frontend-platform/i18n';
+import { useIntl } from '@edx/frontend-platform/i18n';
 
 import { useLibraryPublishHistoryEntries } from '@src/library-authoring/data/apiHooks';
 import { LoadingSpinner } from '@src/generic/Loading';
 
-import { LibraryHistoryEntry, LibraryPublishContributor, LibraryPublishHistoryGroup } from '../../data/api';
+import { LibraryHistoryEntry, LibraryPublishHistoryGroup } from '../../data/api';
 import messages from './messages';
 import { getItemIcon } from '@src/generic/block-type-utils';
-
-const MAX_VISIBLE_CONTRIBUTORS = 5;
+import { ContributorsAvatars } from './ContributorAvatars';
+import { HistoryLogGroupEntries } from './HistoryLogGroupEntries';
 
 export interface HistoryLogGroupTitleProps {
   titleMessage: string | ReactNode;
@@ -34,46 +34,12 @@ export interface HistoryDraftLogGroupProps {
   entries: LibraryHistoryEntry[];
 }
 
-export interface HistoryLogGroupEntriesProps {
-  entries: LibraryHistoryEntry[];
-}
-
 export interface HistoryPublishLogGroupProps extends LibraryPublishHistoryGroup {
   itemId: string;
   // When true, hides the vertical connector line rendered below this group. Used for the last group
   // in the history log to avoid a dangling connector with nothing below it.
   hideLogVert?: boolean;
 }
-
-interface ContributorAvatarProps {
-  username?: string;
-  src?: string;
-  className: string;
-  size: ComponentProps<typeof Avatar>['size'];
-}
-
-interface ContributorsAvatarsProps {
-  contributors: LibraryPublishContributor[];
-}
-
-const ContributorAvatar = ({
-  username,
-  src,
-  className,
-  size,
-}: ContributorAvatarProps) => {
-  const intl = useIntl();
-  const [imgError, setImgError] = useState(false);
-  return (
-    <Avatar
-      className={className}
-      size={size}
-      src={imgError ? undefined : src}
-      alt={username || intl.formatMessage(messages.historyEntryDefaultUser)}
-      onError={() => setImgError(true)}
-    />
-  );
-};
 
 const HistoryLogGroupTitle = ({
   titleMessage,
@@ -101,64 +67,6 @@ const HistoryLogGroupTitle = ({
           </Collapsible.Visible>
         </>
       )}
-    </Stack>
-  );
-};
-
-const HistoryLogGroupEntries = ({
-  entries,
-}: HistoryLogGroupEntriesProps) => {
-  const intl = useIntl();
-
-  const getEntryMessage = (entry: LibraryHistoryEntry) => {
-    switch (entry.action) {
-      case 'edited':
-        return messages.historyEditEntry;
-      case 'renamed':
-        return messages.historyRenameEntry;
-      case 'created':
-        return messages.historyCreatedEntry;
-      default:
-        return messages.historyEditEntry;
-    }
-  };
-
-  return (
-    <Stack gap={0}>
-      <div className="history-log-vert" />
-      {entries.map((entry, index, arr) => {
-        const isLast = index === arr.length - 1;
-        const entryMessage = getEntryMessage(entry);
-
-        return (
-          <div key={entry.changedAt}>
-            <Stack direction="horizontal" gap={2} className="ml-1.5">
-              <ContributorAvatar
-                username={entry.contributor?.username || intl.formatMessage(messages.historyEntryDefaultUser)}
-                src={entry.contributor?.profileImageUrls.medium}
-                className="history-log-group-avatar small-avatar"
-                size="sm"
-              />
-              <Stack>
-                <Stack direction="horizontal" gap={1}>
-                  <FormattedMessage
-                    {...entryMessage}
-                    values={{
-                      user: entry.contributor?.username ?? intl.formatMessage(messages.historyEntryDefaultUser),
-                      displayName: <span className="history-log-title text-truncate">{entry.title}</span>,
-                      icon: <Icon src={getItemIcon(entry.itemType)} />,
-                    }}
-                  />
-                </Stack>
-                <span className="small text-gray-500">
-                  {moment(entry.changedAt).fromNow()}
-                </span>
-              </Stack>
-            </Stack>
-            {!isLast && <div className="history-log-vert" />}
-          </div>
-        );
-      })}
     </Stack>
   );
 };
@@ -220,31 +128,6 @@ export const HistoryDraftLogGroup = ({
       </Collapsible.Advanced>
       <div className="history-log-vert" />
     </div>
-  );
-};
-
-const ContributorsAvatars = ({ contributors }: ContributorsAvatarsProps) => {
-  const visible = contributors.slice(0, MAX_VISIBLE_CONTRIBUTORS);
-  return (
-    <Stack direction="horizontal" gap={2} className="ml-4.5">
-      <div className="contributors-avatars">
-        {visible.map(({ username, profileImageUrls }) => (
-          <ContributorAvatar
-            key={username}
-            size="xs"
-            className="contributors-avatar"
-            username={username}
-            src={profileImageUrls.small}
-          />
-        ))}
-      </div>
-      <span className="small">
-        <FormattedMessage
-          {...messages.historyContributors}
-          values={{ count: contributors.length }}
-        />
-      </span>
-    </Stack>
   );
 };
